+++ README.md
@@ -1,4 +1,4 @@
-wampy.js
+index.js
 ========
 
 Simple WAMP (WebSocket Application Messaging Protocol) Javascript client-side implementation
@@ -102,7 +102,7 @@ Installation
 Wampy.js can be installed using bower or npm or just by file-copy :)
 
 ```bash
-> bower install wampy.js
+> bower install index.js
 # Or
 > npm install wampy
 ```
