+++ src/index.js
@@ -1,0 +1,37 @@
+/**
+ * Project: wampy.js
+ *
+ * https://github.com/KSDaemon/wampy.js
+ *
+ * A lightweight client-side implementation of
+ * WAMP (The WebSocket Application Messaging Protocol v2)
+ * http://wamp.ws
+ *
+ * Provides asynchronous RPC/PubSub over WebSocket.
+ *
+ * Copyright 2014 KSDaemon. Licensed under the MIT License.
+ * See @license text at http://www.opensource.org/licenses/mit-license.php
+ *
+ */
+
+import { Wampy } from './wampy';
+
+'use strict';
+
+// Module boilerplate to support browser globals and browserify and AMD.
+(function (root, m) {
+    if (typeof define === 'function' && define.amd) {
+        // AMD. Register as an anonymous module.
+        define(['exports'], m);
+    } else if (typeof exports === 'object' && typeof exports.nodeName !== 'string') {
+        // CommonJS
+        module.exports = m();
+    } else {
+        // Browser globals
+        root.Wampy = m();
+    }
+}(this, function () {
+
+    return Wampy;
+
+}));
